[C] Use the shared aeron_clock_gettime_monotonic and aeron_clock_gettime_realtime
