+++ aeron-client/src/main/c/aeron_windows.h
@@ -18,6 +18,7 @@
 #define AERON_WINDOWS_H
 
 #include "util/aeron_platform.h"
+#include <stdint.h>
 
 #if defined(AERON_COMPILER_GCC)
 
@@ -28,14 +29,11 @@
 
 #elif defined(AERON_COMPILER_MSVC)
 
-#include <basetsd.h>
 #include <stddef.h>
 
 double aeron_erand48(unsigned short xsubi[3]);
-void aeron_srand48(UINT64 aeron_nano_clock);
+void aeron_srand48(uint64_t aeron_nano_clock);
 double aeron_drand48();
-int aeron_clock_gettime_monotonic(struct timespec *tp);
-int aeron_clock_gettime_realtime(struct timespec *tp);
 char *aeron_strndup(const char *value, size_t length);
 void localtime_r(const time_t *timep, struct tm *result);
